Fixed an issue where we weren't suppressing a call to getCharacterBoundsOnScreen during a scope highlight update.
Previously, we were using a passed parameter to keep track of whether we were during a RichTextFX update (when we must not ask for character positions), but we hadn't covered all cases.  This commit switches to using a field for the flag, set from MoeSyntaxDocument when the document updates begin and end.

(Also got rid of an old inNotification flag on the document, which was fulfilling the same purpose on the document, but was unused.  We now have a used flag in the syntax view, which is the part that cares whether we are in an update.)