Fix BLUEJ-1076: Printing via Stride menu doesn't do anything.
Due to a change for Java printing (I think in commit 06c68de), the Stride behaviour of passing a null print job to printTo became unsupported.  This is now fixed and Stride correctly creates a print job, although printing still only uses a single page (which is a separate bug to this one; see BLUEJ-633).