Fix two issues: the scope backgrounds were not lined up with the printed text, and the right-hand border of the scopes wasn't visible.
The main issue was that we were still using a copy of the on-screen editor's scope backgrounds.  This was fine in 4.1.1 when we were printing using the on-screen editor's font size, but fell apart in 4.1.2 where the printing font size can be different.  So we must recalculate the scopes for the off-screen version.  This turns out to be very problematic because the only API RichTextFX offers is getCharacterBoundsOnScreen, which doesn't work for an off-screen pane.  So we fall back to guessing character positions using our own measureString function, which for some reason requires a slight fudge-factor to look right (see code comments).