Fix BLUEJ-1108: Dependencies are not calculated on first project load.
This is maybe a corner case, since the dependencies are normally loaded
from the package.bluej file. However, these might be out of date or not
present for whatever reason (our people2 project doesn't have
dependencies in the package file file some reason).

The solution is to avoid analysing the source on ClassTarget load, which
was added incorrectly as part of the fix to BLUEJ-508. Doing so at that
stage causes dependency resolution to fail (since the package is "empty"
before the class targets are loaded). By removing this, the later
analysis of the source correctly resolves intrapackage dependencies.