Fix BLUEJ-1111: class type parameters not shown in class diagram.
This got broken at some point between 3.1.7 and 4.0 when the logic for
displaying the ClassTarget label was changed. The "setDisplayName"
setter had then been overridden to append type parameters to the label,
but without being called appropriately when the type parameters changed
nor when they were initially determined.

Since in general overriding a simple setter seems like a bad idea, 
I've refactored the code to update the display name in a separate method
and call it at appropriate junctions.