Improve efficiency of virtual flow listener.
The listener is used to determine when lines are visible and which can
have their scope indents calculated. To do this it was iterating through
every paragraph in the document. This change estimates the visible
paragraph positions instead (shown to be accurate by testing) and falls
back to the old linear search if that fails.