Removed the PackageEditorEvent/PackageEditorListener system.
There was only one listener (PkgMgrFrame) so all the code did was take a bunch of parameters, put them into an over-complicated event object so that they needed casting to pull some of them back out, then call a method on the listener which unpacked the event object and acted on it.  The event and listener are now removed in favour of just directly calling the methods on PkgMgrFrame.

This has bothered me for a while, and making it simpler will make clearer the necessary modifications for the Greenfoot rewrite.