Added a reference to PkgMgrFrame to Package and made the PkgMgrFrame callMethod function route via Package not PackageEditor.
I know that in general we have avoided Package containing a reference to PkgMgrFrame, and instead we use PackageEditor for similar functions.  But something has to give: Greenfoot doesn't create a PackageEditor but now it will need to make a server-side call to methods and constructors.  So either we need to create a PackageEditor in Greenfoot, or use PkgMgrFrame directly from Package.  Since the callMethod call in PackageEditor was only calling PkgMgrFrame anyway, using the PkgMgrFrame reference directly seemed most sensible.

(We could introduce an interface that Package keeps a reference to, which PkgMgrFrame implements, to slightly decouple them, but I don't see there's much gain for us: we won't ever have another implementation of the interface, so it would serve no practical purpose and only be there to make it look like Package doesn't keep a PkgMgrFrame reference.)