Fixed odd API in the editor for inserting code (used by Greenfoot's save the world).
After an old refactoring which moved code around, BlueJ's Editor class had some methods which took as a parameter the extension-wrapped version of the Editor instance: effectively the methods were taking "this" as a parameter.  The current commit removes the parameter in favour of just calling the appropriate unwrapped methods on the current instance.  The commit shouldn't alter any behaviour, but just tidy up the code.