Fixed actor naming and code clearing for save the world.
Previously, actor naming was done by making a direct method call from the debug VM which went to the server VM via RMI.  Now, it uses a special breakpoint for the server VM to spot that naming is needed, and perform the naming.  We will need to check the performance for larger worlds (we may need to batch up the calls), but for now it is working, and that's a good start.  Also fixed when the code for the prepare method and object name cache gets cleared, as it wasn't quite right.