Fix Greenfoot so that user code is run on the simulation thread, not the BlueJ worker thread.
This has to be done in a way that does not make BlueJ parts (e.g. ExecServer) depend on Greenfoot code.  As it happens, we can just extend the recently implemented run-on-thread mechanism in BlueJ (for FX/Swing) to allow a custom runner which Greenfoot sets to use the Simulation thread.