Fix exception bug when pasting class with incorrect package line.
There was a complicated series of events where pasting triggered a save
which triggers a full parse and a package line check (which checks if
the "package xyz;" line matches the package containing the source) and
optional fix. This could result in the document being edited (package
line updated) while in a change event listener, resulting in an
exception.

Ultimately I do not think we should be triggering full parses on saves
due to small edits, and of course ideally we would not save in these
cases but compile from the in-memory source instead. However, that's for
another day. This fix just does some re-arranging and avoids calling
save() from contexts which might be entered via the document change
callback.