Don't maintain paragraph styles separately to the document.
MoeEditorPane was keeping its own map of paragraph styles and style
listeners (labels). This is unnecessary; the styles are also part of the
document, and changing a document paragraph style automatically
re-creates the graphic for that paragraph (as it should).

By using the paragraph styles maintained by the document, MoeEditorPane
is simplified and a bounded memory leak is removed (lines at the end of
the document which were removed wouldn't be removed from the style
maps).