Avoid resetting paragraph style if scope margins unchanged.
The calculation of scope margins caused an unconditional setting of the
paragraph style for the affected lines, even if the scope margins hadn't
changed. This could lead to a cycle, resulting in 100% CPU usage.

Fix is to check that the scope margins have actually changed before
applying the new margins, but this commit also includes some refactoring
which simplifies the code and should make it generally more efficient.