Start to split out BlueJ/Greenfoot GUI functionality into an interface.
Since Greenfoot will be running in the foreground VM, the BlueJ GUI
should no longer be required. Ideally there will no longer need to be a
hidden PkgMgrFrame for every project, etc. This begins the work of
splitting out the UIs into an interface with separate implementations
for BlueJ and Greenfoot.