Fully remove support for non-screen-menubar on Mac.
The bluej.macos.screenmenubar property historically could be used to
disable the screen menu bar on Mac (resulting in the menu bar appearing
at the top of each window instead). Since BlueJ 4.x this has not been
honoured anyway, however the setting still affected the construction of
menus (which it obviously shouldn't).