Force repaint after world changing, so constantly changing world paints.
If the world changes every frame, the runeOneLoop method returns early,
without repainting. Since repainting is when the remote communication
occurs, a scenario that changes the world every act cycle cannot be
paused (and doesn't repaint on the main VM).

(This should be a pretty rare use case, but it is trivial to handle
correctly).