Made the terminal show and come to the front when the simulation stops with an error.
The problem was that the simulation could stop because of an error (especially things like null pointer exceptions, or interacting with an actor after it got removed from the world), but the user may not notice because the terminal would not come to the front.  (It would if it was previously closed, but if the user left the terminal open but behind, they may not see the new output.)  This caused the user to not realise why the simulation had suddenly halted.

There was no single nice solution to this, because it involves a VM hop whatever we do.  Rather than use a breakpoint, I used the intra-VM SHM communication.  Because we don't have a full-on message acknowledgment system in the debug->server VM direction, I send a count each frame, and the server VM watches out for an increase in this count and then shows the terminal and brings it to the front.

Fixes GREENFOOT-375.