Started adding a few thread tags to Greenfoot.
The thread checker is nowhere near passing in Greenfoot, so I haven't actually committed the change to turn it on permanently.  But we can start adding the tags in the mean time without causing any problems in the build, and it's probably going to take several commits over a period of time to add them all, so better to do it gradually than try and it do it in one big lump.