Changed when we open and setup a project on the debug VM.  Previously, we started the simulation thread, then later opened the project.  This was reasonable because the project could change during the debug VM's lifetime.  However, that no longer needs to be true, since the server VM is in charge of opening projects, so we can assume that each debug VM will correspond to a single project.
With the old system, the simulation thread would be started before the world handler delegate was setup, but this meant that early instructions to the simulation (such as instantiate-world) could be ignored, and meant that on project load we weren't always instantiating the world if no compile was needed.  With the new system, the world gets instantiated every time if the classes are already compiled.