Editor: remove catch of exception without logging.
We were catching FileNotFoundException without logging any error. This
was code given to us by a third party a long time ago, who had adapted
the editor as a standalone application and fed us back some of the
changes - however, this particular change is not desirable for us; we
don't need to show an empty editor if the file doesn't exist, and we
want to log the reason for not being able to open the file.