Avoid skipping paints on server, by sending them until ack'd.
There was an issue with some world paints getting lost, since on the
debug VM it could potentially send frames without an image without a
previous (with-image) frame having been received on the other end.

This change keeps the image in the frame until a suitable
acknowledgement (in the form of a sequence number) is received.