Fix logic for greying out world.
Previous logic was to grey the world out if any pending "discard world"
commands had been acknowledged. However this led to race conditions, eg
a new world might have been instantiated since the discard request (in
which case it would incorrectly appear greyed out) or the old world
image might be received after the discard (because the discard is
processed in a runLater in the debugVM), causing the world to display
normally when it should be greyed out.