Revert changes to update handling.
This reverts some changes that were made to update handling which were
not correct or unnecessary. The MoeSyntaxEvent must be passed to the
parseNode.textInsert()/textRemoved() functions to record node structure
changes, which are then used for re-painting.

It's not necessary to call recalculateScopesForLinesInRange() since that
will be done in BlueJSyntaxView.updateDamage() method.