Fixed a race hazard when initialising Greenfoot.
The order of initialisation is that we set the debug VM going, and then we add a debugger listener to watch for the debug VM being ready.  Once we see the ready event, we run some code to launch Greenfoot on the debug VM.  But if too long passes between setting the debug VM going and adding the listener, the ready event may already have occurred.  This was happening because we were creating and showing GreenfootStage in the middle.  But just moving out the GreenfootStage initialisation would still leave a race hazard, just one that is less likely.  The correct fix, done here, is to check when we add the listener if the debug VM is already ready, and if so, launch Greenfoot immediately.