Track lines which require style span updates separately from scopes.
In some cases we have an edit that marks a scope change, but the scopes
do not actually change. In that case we still need to re-do the styles.
In other cases a scope may need to be re-drawn, but we only need to
update the style spans on a smaller number of lines.