Don't overwrite the image in the comms buffer before it's consumed.
We don't want the debug VM writing the image while it is being read by
the server VM, and that is now done with the file lock for the area
held, so we need an alternate synchronisation mechanism to ensure this
doesn't happen. Therefore, have the server VM transmit the sequence of
the last frame it consumed.