Fix a NullPointerException seen during testing.
An NPE can occur when a reset is issued before the simulation thread has
been identified (simulationThread == null in GreenfootDebugHandler). In
particular this occurs if the world class is in uncompiled state when
Greenfoot starts; the world can finish compiling, which triggers a
reset, before the debug VM has even started.