Move WorldCanvas reference from WorldHandler into WorldHandlerDelegateIDE.
Really, WorldCanvas is now DrawVMRemote (and will be renamed something like that in a future change).  The standalone version won't use this, and will reference the inner WorldDisplay which actually displays the world.  Only the IDE version needs this class.  Hence we push it down from WorldHandler into WorldHandlerDelegateIDE, by moving some functionality around, and introduced a couple of methods to WorldHandlerDelegate which WorldHandler used to call directly on WorldCanvas, but which will now have alternate implementations in WorldHandlerDelegateStandAlone (paint is currently left empty, that will be in a future change).