Started changing the standalone version across to use JavaFX.
The change is not yet complete (and GreenfootScenarioViewer and GreenfootScenarioMain are currently a bit messy), but it at least compiles and runs so this seems like a good intermediate step for a merge request.

As part of the change, some old border code got inlined from GreenfootScenarioViewer into Exporter.  I think we still need this (if we do remove it, that will be a separate change).