Perform pick operation for dragging on mouse pressed.
Using the DRAG_DETECTED event is problematic because it is not issued
until the mouse cursor has already moved a certain distance, which may
be larger than the actor (if the actor is small, such as an Ant in the
Ants scenario). In that case the cursor has moved off the actor by the
time the pick request is issued and the actor cannot be dragged.