Removed the notion of triggered listeners.
A triggered listener was one that could start and stop listening, but it is not a concept we are using after the GUI rewrite.  This removes TriggeredListener (whose methods were not being called any more) and all its subclasses.  Also removed the setKeyListener in WorldHandlerDelegate which is not being used any more.