Changed how we pass some properties to the debug VM.
Previously, we used methods in the RMI classes to get paths for the system lib dir and user config dir, as well as a few properties for Config.  This change passes the first two as launch parameters to the simulation VM, and writes all the properties which I could find that the simulation VM reads into a properties file, the path of which is passed so that the properties file can be read on the debug side.

I'm not sure the properties we do put in that file are necessary, hence the catch blocks just proceed anyway even if there's an issue with that properties file.