Fixed a bug where a phantom ask request could be seen.
If the debug VM writes to the shared memory early on, it writes a zero for the ask ID, and the main VM (which had not yet seen any commands answered) would interpret a zero as a valid ask request.  This change requires that the ID be greater than zero for it to count.