getLeftEdge(): fix missed adding of paragraph margin in some cases.
If using a cached or calculated space width, the return value needs to
account for the paragraph graphic margin. Also the calculation of space
widths based on a smaller number of cached space widths assumed that the
editorPane does not add its own margin, which seems risky, so correct
for that here.