Added a new BuiltInGClassNode subclass.
The previous organisation had ended up a bit strange: we used a GClassNode constructor and the setupClassDisplay implementation of GClassNode for Actor/World, then a different constructor and a complete override (no call to super method) in the two subclasses.  This adds a third subclass, pushing all the Actor/World-only parts down into that subclass, which is a better arrangement.