Greenfoot auto-compile: compile class and not whole package.
The current behaviour of compiling the whole package leads to error-free
classes not being marked compiled if there is an error in any other
class. This is potentially quite confusing to the user and in any case
is not helpful, and so is changed by this commit to just compile the
current class (as in BlueJ). Compiling the whole package on an explicit
compile (clicking compile button etc) however is retained.