For explicit compile, ensure selected class is always compiled.
In Greenfoot "compile" in the editor compiles the whole package, but we
should make sure that the selected (edited) class itself is always
compiled.

This is needed especially to make sure that the editor sees the
compilation finish, otherwise it believes that it has a compilation
pending and won't schedule further compilations.