Remove some unneeded code from the thread checker which was causing unnecessary tagging failures.
This code was intended to automatically take the tagged type of an anonymous inner class's parent class and apply it to the inner class so that you didn't have to tag the inner class manually.  But (a) I'm not sure the thinking here was correct, (b) it seems we don't rely on this anywhere in our code, (c) we generally always use lambdas unless we're setting out to tag the class differently, and (d) it clearly caused a false failure in a recent branch.  So now it has been removed.