Avoid use of onceInScene for node that may already be in scene.
The previous commit makes onceInScene work only for nodes not already in
a scene, which was not true for this one case.

The alternatives would be to have onceInScene use "runNowOrLater" or
inhibit the thread checker for the method. Either seems like a kludge to
me (although they should in practice, at least for this particular case,
actually be safe).