Generate Java if necessary on any save.
Some parts of the code rely on the Java source being available after
calling the saveFX() method, so have it generate the Java if it hasn't
yet been done.

Also call setSaved() (notify watcher of save) whenever a save is done,
not just when save() is called. This ensures that class role etc will be
updated consistently, and that isModified() returns a result consistent
with whether setSaved() has been called.