Fix how we number actors for the recording for the prepare() method.
During the rewrite we accidentally changed numbering so that we would name the actors crab, crab1, crab2, unlike the old scheme which went crab, crab2, crab3 (no crab1).  It's important to be consistent (including across Greenfoot versions) because we rely on the consistency in naming when doing multiple save-the-worlds, to ensure names don't overlap.  This commit means saving the world in 3.5.0 which was saved previously in 3.1.0 will not have naming clashes.