Track simulation running/paused state independently of UI state.
When we go from NO_WORLD to having a world, we need to know whether we
are RUNNING or PAUSED, so track running-vs-paused in a separate boolean
and use it to decide which of the two states to transition to.

Also simplify the logic in ControlPanel: we don't need a separate
mechanism to disable all buttons during reset, because we can rely on
the state alone.