Fixed an issue where if you created a world by right-clicking it and invoking the constructor, actors were not being named for save-the-world.
The problem was that the WorldHandlerDelegateIDE didn't have the right class name set for the world being initialised, so it wouldn't spot the actors being named.  Also, the world initialising flag was not set back to false, which this commit also fixes.  (We set it to false when we return to the set the world after the successful construction.)