Change how images are set in the tab headers for editors.
Previously, we used an observable property, fetched via the ClassIconFetcher interface.  But given how we implement the class diagram, this makes fixing the icons on rename awkward, and in retrospect, this was all a bit too awkward.  This commit simplifies the process: instead of making editors fetch their image from the project, we get Greenfoot's class diagram (which already holds a reference to ClassTarget, from which it can access the editor) to set the image on the editor, and introduce a method by which the class diagram knows when an editor has been opened.

This is simpler code now, I think, and fixes most of the issues around editor header icons on rename (where the editor could end up bound to an old property for the previous name).