Fix button types of dialogs.
This adjusts two variants of the askQueiontFX method. The first is used
with a variety of prompts and with buttons in various orders, so we
cannot associate the first button with a "yes" response and so on.

The second is only used once (in Greenfoot), so define a standard
ordering for the buttons.