Various threading fixes.
Again we need to avoid running code on the remote VM synchronously in
the UI event handling thread.

In this commit I've restored the correct return type to
Debugger.instantiateClass(), which returns a (DebuggerObject) result
directly. Having it return an FXPlatformSupplier was actually harmful -
it implied that the debugger took care of obtaining the result in the
background, but it doesn't. In any case, there's no benefit to wrapping
the result in yet another object.

There's a similar issue with many other methods in the Debugger
interface, but they can wait for now.