Fixed how we notice that discarding the world has succeeded.
Previously we waited for a frame without a world, but this may not happen if the world is immediately instantiated again afterwards.  So instead, we now use a counter which is incremented on simulation-side world changes, and when we see this change on the main GUI side, we count the discard as complete.

Also fixed an issue where the simulation could send a stale frame from the double-buffer after a world discard.

Fixes GREENFOOT-780.