Fixed some code which was run on open-project but which should only be run once per window.
This splits the old method into two parts.  The mouse and key handlers are added once, from the constructor, but all have a check for the project being closed.  The animation timer is set off when a project is loaded, but is now also stopped when the project is closed (and is now held in a field for this purpose).