Fix/improve an incorrect message when an update conflict occurs.
There are two basic cases: one where a "CheckoutConflictException"
occurs and we use the (not particularly helpful) message from it
directly, and another case where we get a merge failure result and
incorrectly display a message claiming that an update is required.

What is actually needed in both cases is a commit.