Re-work git update "force files" handling.
The idea is that forced files are "force updated", that is, the
repository version is taken over "our" version. This was accomplished
previously by renaming the file before the update.

There were several bugs here: most obviously, the "package.bluej" file
had special handling and was the only "force file" that was processed.
Only one single package.bluej file was processed (there can be multiple
in a project). Also the file was not restored if the update failed.

This change forces update by checking out the file fresh before the
update (i.e. losing any local changes). This fixes the issues described
above.