Fix an issue where closing the Java editor would not trigger a compilation.
The problem was that deselecting an editor tab (which happens during closing) cancels the fresh (modified) state, which triggers a compilation if there is an unsaved changed -- but some earlier code in the closing process was saving changes.  So we cancel the fresh state before saving, which will correctly trigger an auto-compile when closing a modified editor.

Fixes BLUEJ-1198