Stopped bringing PkgMgrFrame to front once the VM is ready on non-Mac operating systems.
This is needed on Mac because the remote VM steals focus if JavaFX is initialised.  It is bad on Windows because it interferes with the menu focus, and Windows doesn't have this issue with the focus stealing.  So we only run this code on Mac, and both work fine.  (Linux doesn't mind, but we also don't run it on Linux.)

Fixes BLUEJ-1200