Fixed the module select dialog when checking out.
It is now a proper FX dialog with a return type rather than using the old Swing-like model.  It also makes the module list more useful -- single clicking a module fills in the text field at the top with the module name, and double-clicking the module list closes the dialog.

(Previously, the module list was displayed, but single or double clicking it did nothing useful.)

Fixes BLUEJ-1202 (problem 2)