Fix NPE in clearAllSelections.
In a project with packages, some of which haven't been opened,
clearAllSelections() will cause a NullPointerException. A similar issue
exists in selectTargetsInGraphs(). These methods are called on a
merge/update if there are conflicts.

This change fixes these issues, and also opens package editors for
packages containing files with conflicts, if they are not already open.