Remove use of FX class (Platform) from appToFront().
Firstly, the need to distinguish FX platform -vs- non-FX platform thread
shouldn't be necessary here. The 'if' statement which does this was
added in a commit during GF3 development but the comment doesn't make
sense; the wait that it avoids doesn't call anything on the FX thread
and doesn't deadlock (tested just now).

This commit just reverts that one - always perform the wait, regardless
of what thread we are running on. Note that this normally is
unnoticeable because the script executes very quickly, and is half a
second at most.