Fix mouse input for worlds with a cell size > 1.
This turns out to be non-trivial since earlier changes discarded the
pixel position information from events and then later determined
associated actor based on the recorded position (which was a cell
position).

However, the pixel position is required, since click detection is pixel
perfect even with large cells (consider the case of an actor spanning
multiple cells but not completely covering cells at the edge. A click
just on the edge of the actor should register, but outside of it should
not).

So, in this commit, store the pixel position together with the cell
position in MouseInfo, and use the pixel position to find associated
actors.