Remove the hack which reduces build output verbosity.
Other than that it uses javascript within an ant build file, which
possibly won't work in the future anyway, it is a terrible idea as it
makes debugging build issues very difficult (especially as build
documentation does not mention it nor how to prevent it) and prevents
the usual ant verbosity controls from working.

The correct way to quieten the output is to run with -q or -S on the ant
command line. In the general case, more output by default is the better
option.