Changed the launcher EXE files to be 64-bit.
Since the JVM is now 64-bit, and we load the jvm.dll directly from the launcher, our EXE must now also be 64-bit.  That requires a 64-bit toolchain, so we must switch to MinGW64.

This commit also removes the msys aspects, in favour of just using Windows binaries for make/g++/etc directly.