Allow specification of the JavaFX path on the command-line, rather than assuming it is bundled.
This fixes two separate situations.  One is that during IDE development, JavaFX is not bundled inside the BlueJ directory, so we need a way to specify its location for the purpose of dynamically building classpaths featuring JavaFX.  The new "-javafxpath" property provides this, and should always be specified when running from the IDE, alongside -useclassesdir and others.

The other problem is that on Linux we probably will not bundle JavaFX, so we need to specify the JavaFX path in the launcher, as a directory rather than the previous jfxrt.jar single-file path.  So we re-use the same javafxpath property and set it in the launcher on Linux.  The default examined path for JavaFX on Linux is a based on the openjfx 11 package in Ubuntu cosmic: https://packages.ubuntu.com/cosmic/amd64/openjfx/filelist and Debian sid: https://packages.debian.org/sid/amd64/openjfx/filelist