Fix a bug where the debugger may show empty if a breakpoint is hit on a system thread while system threads are hidden.
The fix is two-fold.  Firstly, it stops counting the AWT and FX event threads as system threads -- user code frequently runs on them (and potentially all user code if using our run-on-AWT or run-on-FX setting), so it seems odd to hide them as being system threads, especially as we don't necessarily initialise FX or AWT if the user code isn't using it to begin with.

The second part of the fix is a backup: if we do try to select a thread that is hidden due to being a system thread, we show system threads to avoid the debugger being empty.  This should be pretty rare with the above change, but worth doing to avoid user confusion.

Fixes BLUEJ-1208