Fix a bug in GifImage where it wasn't overwriting the previous image region correctly.
As described on the Greenroom:  The bug was that the colour with which to overwrite the rectangle may be totally transparent. If you ask to fill a rectangle with transparency, it actually blends it over the top, thus not clearing the previous image. The previous GifImage used a different alpha blending mode to avoid this issue, but since we don’t expose that in GreenfootImage, I’ve instead used a loop that directly writes the colour value to each pixel.