Fixed a problem where adding an item to the method body at a low indent could move the indent of the outer class body, which is different to the original scope highlighting, and also was not changing back on a later edit.
The problem was that if you have a class with a method in it, then when you alter the method, the parse node for the method body is removed pending a reparse.  While it is removed, the calculation for the class/method indent will count the method body content against its indent and move the indent outwards.  But once the inner body was re-added as a node, nothing was recalculating the indent of the outer nodes.

We fix this using a two-pronged approach.  One is that when an inner node is changed or removed, we remove the cached indent for parent nodes.  That way, we mark that recalculation is needed.  The second is that we form an association between scoped backgrounds and the node they come from.  When the indent is re-cached for the parent nodes, we can update the graphical scopes without needing to reparse the entire content of the node just to fix up that scope (which probably changed just due to one line, which itself has already been reprocessed).

With these changes, testScope2 now passes, and the behaviour in the editor generally seems good, also with no scope flickering.