Add double-buffering to the Greenfoot GUI side.
This doesn't fix the flickering issue with the world display, but while investigating that issue, I realised there is the potential for problems here.  Although we do it on the FX thread, we are writing for a prolonged period to an image which is currently live on screen (by being held in the ImageView) and potentially notifying listeners, so it seems neater to write to an image off-screen, then swap it in.