Fix a bug where Greenfoot could render into an image at the same time as it was reading the data to send to the other VM.
I never liked the original code as it was difficult to reason about.  This switches to using two concurrency primitives: there's an atomic reference to the image waiting to be sent (always flows from rendering thread to comms thread), and a queue of old images to re-use (which flows from the comms thread back to the rendering thread).  This way, images are always held by one thread or the other as they are sent around, never both.

Fixes GREENFOOT-836.