Rewrote VMEventHandler's threading.
Previously, it was blocking on the JDI event queue, then using a complicated system of calling interrupt() to inject other events on to the thread.  This change adds a further thread that is just dedicated to copying JDI events on to a shared event queue, which is then blocked on by the VMEventHandler thread.  Other threads can then easily add events to this queue.