Fixed a threading issue with resuming the server thread which could cause a NullPointerException.
The issue was that the thread (which lives on the debug VM) could be resumed, finished, and then we go looking for its internal reference in the debugger.  But we can still be interrogating the JdiThread representation of it on the server VM side so we don't want to just ignore it.  The solution is to hold a reference to the thread permanently as a JdiThread rather than ThreadReference, which means we never have to go looking for it in the the thread set later.