Guarded against Errors being thrown in extensions.
This old code is intended to prevent problems in extensions from propagating back into BlueJ by catching all Exceptions.  But some problems cause an Error to be thrown; Error is a subclass of Throwable but not of Exception.  This is relevant for BlueJ 5 because, as I noticed earlier, old extensions are likely to throw NoSuchFieldError while looking for the extension version field, which can prevent BlueJ loading.

(I'm not fussed that we broke the old extensions in this way, because they are bound to be broken in one way or another, we just don't want it to prevent BlueJ loading)