Fixed a bug where sometimes the positions of the scope backgrounds would not be calculated correctly.
It seems that sometimes the font CSS will not yet have been applied to the Text items, and if we try to calculate the left edge of the text at this time, we will get the wrong answer, and cache it for later.  So we must manually check the font before calculating the left edge, to make sure the font is correct.