Fixed a bug where Windows line endings could get pasted into the editor.
They were already removed when loading a file, and could not be typed, but pasting allowed them in.  This would then cause problems with, for example, the End key, which would try to place the caret between the '\r' and the '\n', making it look like the caret had jumped to the beginning of the next line.