Fixed a bug where our dependency checking for compilation was not checking all indirect dependencies thoroughly.  It would check the direct dependencies, and their parents, but not the dependencies' dependencies.
Because of this, there was an infinite loop bug as follows.  Class A had an error, and B depended on A and C depended on B.  We tried to compile all classes and it failed because of A.  We then tried to independently compile any classes not dependent on A.  When we analysed the dependencies for C, we found it depended on B, but we didn't see that B depended on A and thus we tried to compile C, which in turn tried to compile A.  But we should not have tried C because it was ultimately dependent on A.

The new algorithm first gathers up all dependencies (taking care to avoid an infinite loop if there was a mutual dependency) then checks them to see if any have errors.