Worked around a JDK bug which was getting in the way of loading Oracle Javadoc.
The bug is https://bugs.openjdk.java.net/browse/JDK-8213202 and in our case it caused a problem that when the JavaFX WebView tried to load Oracle Javadoc, it would fail to load many of the assets (e.g. CSS) and so would show a badly formatted page.  This is fixed in JDK 12, but since we're on 11 for the time being, we need to work around it.  The work around suggested in the bug report seems to work so I've added it to all our various launchers.  See also https://stackoverflow.com/questions/52574050/javax-net-ssl-sslexception-no-psk-available-unable-to-resume which pointed me to the bug and suggested another fix (which we didn't need).