Adjusted what happens when a class target is invalidated.
Previously, it was marked as modified, which means it only changed state if it was previously compiled.  Initially this seems reasonable -- if B depends on A and A is modified, then even if B is not modified, B must now be marked as needing compilation.  But this is not quite right -- if a class B has errors, and class A changes, those errors in B may no longer be valid either, so we still need to mark B as needing compilation, rather than has-errors.  Essentially, compiled and has-errors are the two definite states, and needs-compile is the unknown state.  As part of this change, we must also remove any compilation errors from class B.